Move screen detector to separated module
